removed some duplicate code caused by replacement of filters with selections
